Move buildable kit types to /models
