+++ src/models/kit/index.ts
@@ -6,11 +6,14 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
+export * from './buildable-kit';
 export * from './is-local-kit';
 export * from './is-remote-kit';
 export * from './local-kit';
+export * from './raw-buildable-kit';
 export * from './raw-local-kit';
 export * from './raw-remote-kit';
 export * from './remote-kit';
 export * from './to-local-kit';
+export * from './to-raw-buildable-kit';
 export * from './to-remote-kit';
